전체조회 : 댓글, 좋아요 갯수 기능 추가완료 ( 단, 조회성능은 떨어짐.. 추후 수정 )

+++ src/main/java/com/example/blogStudy/repository/LikeRepository.java
@@ -4,7 +4,7 @@ import com.example.blogStudy.entity.Like;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LikeRepository extends JpaRepository<Like, Long> {
-    int countByPostId(Long postId);
+    long countByPostId(Long postId);
 
     boolean existsByUserIdAndPostId(String id, Long postId);
 }
